created seperate components, updated media queries
