+++ src/styles/App.scss
@@ -15,6 +15,7 @@ html {
 body {
     background-color: $color-green-transparent;
     font-family: $font-body;
+    color: $color-brown;
     @include fluidFontSize(1.7rem, 2rem);
 }
 
@@ -34,38 +35,36 @@ body > div {
 main {
     display: flex;
     flex-direction: column;
-    margin: 0 10px;
-    position: relative;
-    top: -20vw;
+    margin-top: 70px;
     width: 100%;
 
     //MQ: center flex items and remove margin
     @include tablet {
         & {
-            margin: 0;
+            // margin: 0;
             align-items: center;
         }
     }
 
-    //MQ: move up on page
-    @include largePhone {
-        & {
-            top: -28vw;
-        }
-    }
+    // //MQ: move up on page
+    // @include largePhone {
+    //     & {
+    //         margin-top: 28vw;
+    //     }
+    // }
 
-    @include smallPhone {
-        & {
-            top: -23vw;
-        }
-    }
+    // @include smallPhone {
+    //     & {
+    //         margin-top: 23vw;
+    //     }
+    // }
 
-    //hard code top position beyond max width
-    @include minWidth {
-        & {
-            top: -245px;
-        }
-    }
+    // //hard code top position beyond max width
+    // @include minWidth {
+    //     & {
+    //         margin-top: 245px;
+    //     }
+    // }
 }
 
 section {
@@ -88,6 +87,7 @@ p {
 a {
     text-decoration: none;
     color: $color-light;
+    font-family: $font-body;
 }
 
 /* ||||||||||| HEADER + BACKGROUND ||||||||||| */
@@ -137,28 +137,12 @@ a {
 }
 
 .headerImg {
-    width: 30%;
+    max-width: 400px;
+    width: 40vw;
     margin-top: 25px;
     margin-left: 5px;
-
-    //MQ: increase size
-    @include tablet {
-        & {
-            width: 40%;
-        }
-    }
-
-    @include largePhone {
-        & {
-            width: 50%;
-        }
-    }
-
-    @include smallPhone {
-        & {
-            width: 60%;
-        }
-    }
+    position: absolute;
+    z-index: -5;
 }
 
 //constructing instructions and reset signs
@@ -169,8 +153,8 @@ a {
     width: unset;
     color: $color-light;
     @include fluidFontSize(1.8rem, 3rem);
-    position: absolute;
     box-shadow: 2px 1px 2px 1px $color-brown-transparent;
+    margin-top: 20px;
 
     div {
         border: solid $color-light 3px;
@@ -179,11 +163,13 @@ a {
         position: relative;
     }
 
-    button {
+    button, a {
         border: none;
         background: transparent;
         color: $color-light;
         font-family: $font-body;
+        font-weight: 500;
+        padding: 0;
     }
 
     .signPole {
@@ -198,8 +184,11 @@ a {
     }
 
     &.instructions {
-        top: 60px;
-        right: 95px;
+        margin-right: 53px;
+    }
+
+    &.path {
+        margin-right: 20px;
     }
 
     &.reset {
@@ -213,6 +202,10 @@ a {
             background-color: $color-light;
             color: $color-brown;
 
+            a {
+                color: $color-brown;
+            }
+
             div {
                 border: solid $color-brown 3px;
                 padding: 5px;
@@ -220,16 +213,14 @@ a {
 
             button {
                 color: $color-brown;
-                font-weight: 500;
             }
         
             .signPole {
                 display: none;
             }
 
-            &.instructions {
-                top: calc(23vw - 10px);
-                right: 5%;
+            &.instructions, &.path {
+                margin-right: 5%;
             }
 
             &.reset {
@@ -247,23 +238,24 @@ a {
             & div {
                 padding: 3px;
             }
-
-            &.instructions {
-                top: calc(29vw - 10px);
-            }
-        }
-    }
-
-    @include smallPhone {
-        &.instructions {
-            top: calc(45vw - 10px);
         }
     }
     
 }
 
-nav li {
-    display: inline-block;
+nav ul {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-end;
+
+
+    li {
+        display: inline-block;
+    }
+
+    a {
+        color: $color-light;
+    }
 }
 
 li .instructions {
@@ -369,7 +361,7 @@ li .instructions {
 
 /* ||||||||||| INPUT + PROGRESS SECTIONS ||||||||||| */
 
-.wordInputSection, .progressSection {
+.wordInputSection, .progressSection, .collectionSection {
     padding: 30px;
     width: 60%;
     background-color: $color-light-transparent;
@@ -485,6 +477,43 @@ input[type="text"]:focus {
     color: $color-brown;
     transform: scale(1.05);
     box-shadow: 1px 1px 3px 1px $color-brown-transparent;
+}
+
+//collection section
+.collectionSection {
+    width: 80%;
+    margin: 0 auto;
+    overflow-y: scroll;
+    display: flex;
+    flex-wrap: wrap;
+
+    h2 {
+        width: 100%;
+    }
+
+    > div {
+        background-color: $color-brown;
+        border-radius: 5px;
+        box-shadow: 1px 1px 3px 1px $color-brown-transparent;
+        color: $color-light;
+        padding: 10px;
+        margin: 10px;
+        width: calc(80% / 3);
+    }
+
+    //scrollbar
+    &::-webkit-scrollbar {
+        width: 20px;
+    }
+
+    &::-webkit-scrollbar-track {
+        background: $color-light;
+    }
+    
+    //handle
+    &::-webkit-scrollbar-thumb {
+        background: $color-brown-transparent;
+    }
 }
 
 /* ||||||||||| FOOTER ||||||||||| */
